WIS-33-delete-user-endpoint Added UserNotFoundException to deleteUser in UserService.java

+++ src/main/java/com/caldev/wishlister/services/UserService.java
@@ -3,8 +3,10 @@ package com.caldev.wishlister.services;
 import com.caldev.wishlister.dtos.NewUserDto;
 import com.caldev.wishlister.entities.Authority;
 import com.caldev.wishlister.entities.UserAccount;
+import com.caldev.wishlister.exceptions.UserNotFoundException;
 import com.caldev.wishlister.repositories.AuthorityRepository;
 import com.caldev.wishlister.repositories.UserManagementRepository;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 
 import java.util.*;
@@ -43,6 +45,9 @@ public class UserService {
     }
 
     public void deleteUser(UUID requestedId){
+        if (userManagementRepository.findById(requestedId).isEmpty()){
+            throw new UserNotFoundException("User not found with id: " + requestedId);
+        }
         userManagementRepository.deleteById(requestedId);
     }
 }
